feat[#365] : useSnackbar 커스텀 훅 생성
